requirements.txt
pytest

project-version
python-dotenv
pyannotate

pydub
torch
pyannote-audio
openai-whisper
transformers
torchaudio
soundfile